--- src/main/java/com/account/AccountApplication.java
@@ -1,13 +1,0 @@
-package com.account;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AccountApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AccountApplication.class, args);
-	}
-
-}
