--- src/main/java/com/account/dto/ResponseDto.java
@@ -1,24 +1,0 @@
-package com.account.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Schema(
-        name = "Response",
-        description = "Schema to hold successful response information"
-)
-@Data @AllArgsConstructor
-public class ResponseDto {
-
-    @Schema(
-            description = "Status code in the response"
-    )
-    private String statusCode;
-
-    @Schema(
-            description = "Status message in the response"
-    )
-    private String statusMsg;
-    
-}
