--- src/main/java/com/account/entity/Accounts.java
@@ -1,23 +1,0 @@
-package com.account.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter @Setter @ToString @AllArgsConstructor @NoArgsConstructor
-public class Accounts extends  BaseEntity {
-
-    @Column(name="customer_id")
-    private Long customerId;
-
-    @Column(name="account_number")
-    @Id
-    private Long accountNumber;
-
-    @Column(name="account_type")
-    private String accountType;
-
-    @Column(name="branch_address")
-    private String branchAddress;
-
-}
