--- src/main/java/com/account/entity/Customer.java
@@ -1,22 +1,0 @@
-package com.account.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter @Setter @ToString @AllArgsConstructor @NoArgsConstructor
-public class Customer extends  BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="customer_id")
-    private Long customerId;
-
-    private String name;
-
-    private String email;
-
-    @Column(name="mobile_number")
-    private String mobileNumber;
-
-}
