--- src/main/java/com/account/exception/ResourceNotFoundException.java
@@ -1,13 +1,0 @@
-package com.account.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String resourceName, String fieldName, String fieldValue) {
-        super(String.format("%s not found with the given input data %s : '%s'", resourceName, fieldName, fieldValue));
-    }
-
-}
