--- src/main/java/com/account/repository/CustomerRepository.java
@@ -1,16 +1,0 @@
-package com.account.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.account.entity.Customer;
-
-import java.util.Optional;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-
-    Optional<Customer> findByMobileNumber(String mobileNumber);
-
-
-}
