--- src/main/java/com/account/service/IAccountsService.java
@@ -1,35 +1,0 @@
-package com.account.service;
-
-import com.account.dto.CustomerDto;
-
-public interface IAccountsService {
-
-    /**
-     *
-     * @param customerDto - CustomerDto Object
-     */
-    void createAccount(CustomerDto customerDto);
-
-    /**
-     *
-     * @param mobileNumber - Input Mobile Number
-     * @return Accounts Details based on a given mobileNumber
-     */
-    CustomerDto fetchAccount(String mobileNumber);
-
-    /**
-     *
-     * @param customerDto - CustomerDto Object
-     * @return boolean indicating if the update of Account details is successful or not
-     */
-    boolean updateAccount(CustomerDto customerDto);
-
-    /**
-     *
-     * @param mobileNumber - Input Mobile Number
-     * @return boolean indicating if the delete of Account details is successful or not
-     */
-    boolean deleteAccount(String mobileNumber);
-
-
-}
