+++ src/main/java/com/us/EComUserServiceApplication.java
@@ -1,0 +1,13 @@
+package com.us;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EComUserServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EComUserServiceApplication.class, args);
+	}
+
+}
