+++ src/main/java/com/us/audit/AuditAwareImpl.java
@@ -1,0 +1,21 @@
+package com.us.audit;
+
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.stereotype.Component;
+
+import java.util.Optional;
+
+@Component("auditAwareImpl")
+public class AuditAwareImpl implements AuditorAware<String> {
+
+    /**
+     * Returns the current auditor of the application.
+     *
+     * @return the current auditor.
+     */
+    @Override
+    public Optional<String> getCurrentAuditor() {
+        return Optional.of("ACCOUNTS_MS");
+    }
+	
+}
