+++ src/main/java/com/us/controller/UserRegController.java
@@ -1,0 +1,12 @@
+package com.us.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+
+@RestController
+@RequestMapping("/api/v1/user")
+public class UserRegController {
+	
+}
