+++ src/main/java/com/us/exception/CustomerAlreadyExistsException.java
@@ -1,0 +1,13 @@
+package com.us.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
+public class CustomerAlreadyExistsException extends RuntimeException {
+
+    public CustomerAlreadyExistsException(String message) {
+        super(message);
+    }
+
+}
