+++ src/main/java/com/us/repository/UserRegRepository.java
@@ -1,0 +1,16 @@
+package com.us.repository;
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import org.springframework.stereotype.Repository;
+
+import com.us.entity.UserReg;
+
+
+@Repository
+public interface UserRegRepository extends JpaRepository<UserReg, Long> {
+	Optional<UserReg> findById(Long id);
+  
+
+}
