+++ src/main/resources/application.yml
@@ -1,14 +1,14 @@
 spring:
   application:
-    name: account
+    name: user
   datasource:
     password: root
-    url: jdbc:mysql://localhost:3306/account
+    url: jdbc:mysql://localhost:3306/user
     username: root
     driver-class-name: com.mysql.cj.jdbc.Driver  
   jpa:
     hibernate:
       ddl-auto: update
 server:
-  port: 9092
+  port: 9091
