+++ src/main/resources/schema.sql
@@ -1,19 +1,20 @@
-CREATE TABLE IF NOT EXISTS `customer` (
-  `customer_id` int AUTO_INCREMENT  PRIMARY KEY,
+CREATE TABLE IF NOT EXISTS `user_reg` (
+  `user_id` int AUTO_INCREMENT  PRIMARY KEY,
   `name` varchar(100) NOT NULL,
   `email` varchar(100) NOT NULL,
   `mobile_number` varchar(20) NOT NULL,
+  `address_id` int NOT NULL,
   `created_at` date NOT NULL,
   `created_by` varchar(20) NOT NULL,
   `updated_at` date DEFAULT NULL,
     `updated_by` varchar(20) DEFAULT NULL
 );
 
-CREATE TABLE IF NOT EXISTS `accounts` (
-  `customer_id` int NOT NULL,
-   `account_number` int AUTO_INCREMENT  PRIMARY KEY,
-  `account_type` varchar(100) NOT NULL,
-  `branch_address` varchar(200) NOT NULL,
+CREATE TABLE IF NOT EXISTS `address` (
+  `address_id` int AUTO_INCREMENT  PRIMARY KEY
+  `city_name` varchar(100) NOT NULL,
+  `state_name` varchar(200) NOT NULL,
+  `pin_code` int NOT NULL,
   `created_at` date NOT NULL,
    `created_by` varchar(20) NOT NULL,
    `updated_at` date DEFAULT NULL,
