--- src/test/java/com/account/AccountApplicationTests.java
@@ -1,13 +1,0 @@
-package com.account;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AccountApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
