+++ src/test/java/com/us/EComUserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.us;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EComUserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
